Adds capability of choosing 'personalized' options (eg, columns)
