+++ metabase_api/utility/personalization_json_schema.json
@@ -1,0 +1,21 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://example.com/weird_product.schema.json",
+  "title": "Insight",
+  "description": "Format of the personalization of a dashboard (at deployment time)",
+  "type": "object",
+  "properties": {
+    "fields_replacements": {
+      "$ref": "#/definitions/fields_replacements"
+    }
+  },
+  "definitions": {
+    "fields_replacements": {
+      "type": "object",
+      "description": "Definitions of replacements for fields"
+    }
+  },
+  "required": [
+    "fields_replacements"
+  ]
+}
